// web/src/api/generated/models/activity.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Kick Sim Studio API
 * OpenAPI spec version: 0.2.0
 */

export interface Activity {
  attemptId: string;
  createdAt: string;
  durationMs: number;
  eventType: string;
  eventVersion: number;
  outcome: string;
  runId: string;
  scenarioDefinitionId?: string;
  status?: number;
}