// web/src/api/generated/models/bootstrap.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Kick Sim Studio API
 * OpenAPI spec version: 0.2.0
 */
import type { Activity } from './activity';
import type { KeyInfo } from './keyInfo';
import type { Workspace } from './workspace';

export interface Bootstrap {
  /** A URL to the JSON Schema for this object. */
  readonly $schema?: string;
  apiVersion: number;
  /** @nullable */
  capabilities: string[] | null;
  key: KeyInfo;
  productVersion: string;
  /** @nullable */
  recentActivity: Activity[] | null;
  workspace: Workspace;
}