// web/src/api/generated/models/deliveryAttemptDetail.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Kick Sim Studio API
 * OpenAPI spec version: 0.2.0
 */
import type { Attempt } from './attempt';
import type { Event } from './event';
import type { Run } from './run';

export interface DeliveryAttemptDetail {
  /** A URL to the JSON Schema for this object. */
  readonly $schema?: string;
  attempt: Attempt;
  event: Event;
  run: Run;
  signatureInput: string;
}