// web/src/api/generated/models/keyInfo.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Kick Sim Studio API
 * OpenAPI spec version: 0.2.0
 */

export interface KeyInfo {
  /** A URL to the JSON Schema for this object. */
  readonly $schema?: string;
  algorithm: string;
  bits: number;
  fingerprint: string;
  matchingPrivateKey: boolean;
  modifiedAt: string;
  path: string;
}